pb: some project update edits
I decided to have add an introduction and bring the goals to the top. I tweaked the wording of somethings and deleted some parts from the research section that didn't add much. I also removed the visualization section and added our visual near the top by the initial data dive in section.